[refac] Rename MemberBatchFacade package for better organization
